simplified: removed ArtifactStore and ProgramStore
